+++ src/main/java/absolutelyaya/ultracraft/client/gui/screen/ServerConfigScreen.java
@@ -66,6 +66,7 @@ public class ServerConfigScreen extends Screen
 		addRule(GameruleRegistry.NAILGUN_DAMAGE, pos, GameRuleWidget.ValueType.INT);
 		addRule(GameruleRegistry.HELL_OBSERVER_INTERVAL, pos, GameRuleWidget.ValueType.INT);
 		addRule(GameruleRegistry.START_WITH_PIERCER, pos, GameRuleWidget.ValueType.BOOL);
+		addRule(GameruleRegistry.BLOOD_SATURATION, pos, GameRuleWidget.ValueType.BOOL);
 		
 		boolean b = false;
 		if(simplistic != null)